+++ src/main/java/com/itmentorcommunityplatform/dataimporter/dto/response/ImportStartResponseDto.java
@@ -1,4 +1,4 @@
-package com.itmentorcommunityplatform.dataimporter.dto;
+package com.itmentorcommunityplatform.dataimporter.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
